Move app logic into app.go
